requirements.txt
pytesseract
pillow
pipenv